#6 Correções;

+++ src/main/java/br/edu/ifpb/dac/entidades/Venda.java
@@ -10,7 +10,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToOne;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 
@@ -23,7 +22,7 @@ import javax.persistence.TemporalType;
 public class Venda implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     
     @Temporal(TemporalType.DATE)
     private Date dataVenda;
@@ -46,11 +45,11 @@ public class Venda implements Serializable {
         this.total = total;        
     }
     
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
